# specs/api_contracts.yaml
openapi: "3.1.0"
info:
  title: ChatBridge v2 API
  version: "1.0.0"
  description: "K-12 AI chat platform with third-party app integration. Fastify backend with auto-generated OpenAPI (L-001)."
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /auth/lti/launch:
    post:
      summary: LTI 1.3 OIDC launch
      description: Handles LTI 1.3 launch from Canvas/Schoology. Extracts role, district, issues JWT.
      tags: [Auth]
      operationId: ltiLaunch
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id_token: { type: string }
                state: { type: string }
      responses:
        "302": { description: "Redirect to frontend with JWT in cookie" }
        "401": { description: "LTI verification failed" }
      x-requirement: FR-AUTH-001
      x-agent-hints: "Handle via ltijs library. Extract role + district_id from LTI claims."
  /auth/login:
    post:
      summary: Platform JWT login
      description: Fallback login when LTI is unavailable.
      tags: [Auth]
      operationId: platformLogin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  role: { type: string, enum: [student, teacher, district_admin] }
        "401": { description: "Invalid credentials" }
      x-requirement: FR-CHAT-008
...
  /auth/oauth/spotify/callback:
    post:
      summary: Spotify OAuth2 callback
      tags: [Auth]
      operationId: spotifyCallback
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                state: { type: string }
      responses:
        "200": { description: "Token stored, popup can close" }
        "400": { description: "OAuth error" }
      x-requirement: FR-AUTH-003
...
  /conversations/{conversationId}/messages:
    post:
      summary: Send a message
      description: Student sends message. Triggers safety pipeline → AI response → optional tool invocation.
      tags: [Chat]
      operationId: sendMessage
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 4000 }
      responses:
        "202": { description: "Accepted for async processing (safe/warning)" }
        "200": { description: "Crisis resources returned (critical severity)" }
        "422": { description: "Message blocked by safety pipeline" }
        "429": { description: "Rate limit exceeded" }
      x-requirement: FR-CHAT-001
      x-agent-hints: "Run 4-stage safety pipeline before AI. Stream response via WebSocket."

    get:
      summary: Get conversation history
      description: Paginated message history including app cards.
      tags: [Chat]
      operationId: getMessages
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: before
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages: { type: array }
                  has_more: { type: boolean }
      x-requirement: FR-CHAT-003

  # === Apps ===
...